requirements.txt
streamlit
selenium
beautifulsoup4
pandas
requests
groq
python-dotenv
chromadb 
langchain  
Numpy
langchain-groq